requirements.txt
Django>=3.0,<4.0
google-cloud-speech
google-cloud-texttospeech
psycopg2-binary
openai